Feature: adds requirements.txt.lock file

+++ install.sh
@@ -2,19 +2,17 @@
 set -e # everything must succeed.
 echo "[-] install.sh"
 
-# requires 3.6+ , depends on dict maintaining order
-python=/usr/bin/python3.6
-py=${python##*/} # ll: python3.6
-
-# check for exact version of python3
-if [ ! -e "venv/bin/$py" ]; then
-    echo "could not find venv/bin/$py, recreating venv"
-    rm -rf venv
-    $python -m venv venv
-fi
+. mkvenv.sh
 
 source venv/bin/activate
 
-pip install -r requirements.txt --no-cache-dir
+if [ ! -e requirements.txt.lock ]; then
+    # initial case
+    pip install -r requirements.txt --no-cache-dir
+    pip freeze > requirements.txt.lock
+else
+    # normal case
+    pip install -r requirements.txt.lock
+fi
 
 echo "[✓] install.sh"
